// @flywave/flywave-gltf/src/gltf/types/gltf-json-schema.d.ts
export type GLTFId = number;
// ...
export interface GLTFAccessorSparse {
    /**
     * Number of entries stored in the sparse array.
     */
    count: number;
    /**
     * Index array of size `count` that points to those accessor attributes that deviate from their initialization value. Indices must strictly increase.
     */
    indices: GLTFAccessorSparseIndices;
    /**
     * Array of size `count` times number of components, storing the displaced accessor attributes pointed by `indices`. Substituted values must have the same `componentType` and number of components as the base accessor.
     */
    values: GLTFAccessorSparseValues;
    extensions?: Record<string, any>;
    extras?: any;
}
/**
 * A typed view into a bufferView.  A bufferView contains raw binary data.  An accessor provides a typed view into a bufferView or a subset of a bufferView similar to how WebGL's `vertexAttribPointer()` defines an attribute in a buffer.
 */
export interface GLTFAccessor {
    /**
     * The index of the bufferView.
     */
    bufferView?: GLTFId;
    /**
     * The offset relative to the start of the bufferView in bytes.
     */
    byteOffset?: number;
    /**
     * The datatype of components in the attribute.
     */
    componentType: 5120 | 5121 | 5122 | 5123 | 5125 | 5126 | number;
    /**
     * Specifies whether integer data values should be normalized.
     */
    normalized?: boolean;
    /**
     * The number of attributes referenced by this accessor.
     */
    count: number;
    /**
     * Specifies if the attribute is a scalar, vector, or matrix.
     */
    type: "SCALAR" | "VEC2" | "VEC3" | "VEC4" | "MAT2" | "MAT3" | "MAT4" | string;
    /**
     * Maximum value of each component in this attribute.
     */
    max?: number[];
    /**
     * Minimum value of each component in this attribute.
     */
    min?: number[];
    /**
     * Sparse storage of attributes that deviate from their initialization value.
     */
    sparse?: GLTFAccessorSparse;
    name?: any;
    extensions?: Record<string, any>;
    extras?: any;
}
// ...
export interface GLTFAnimationChannel {
    /**
     * The index of a sampler in this animation used to compute the value for the target.
     */
    sampler: GLTFId;
    /**
     * The index of the node and TRS property to target.
     */
    target: GLTFAnimationChannelTarget;
    extensions?: Record<string, any>;
    extras?: any;
}
/**
 * Combines input and output accessors with an interpolation algorithm to define a keyframe graph (but not its target).
 */
export interface GLTFAnimationSampler {
    /**
     * The index of an accessor containing keyframe input values, e.g., time.
     */
    input: GLTFId;
    /**
     * Interpolation algorithm.
     */
    interpolation?: "LINEAR" | "STEP" | "CUBICSPLINE" | string;
    /**
     * The index of an accessor, containing keyframe output values.
     */
    output: GLTFId;
    extensions?: Record<string, any>;
    extras?: any;
}
/**
 * A keyframe animation.
 */
export interface GLTFAnimation {
    /**
     * An array of channels, each of which targets an animation's sampler at a node's property. Different channels of the same animation can't have equal targets.
     */
    channels: GLTFAnimationChannel[];
    /**
     * An array of samplers that combines input and output accessors with an interpolation algorithm to define a keyframe graph (but not its target).
     */
    samplers: GLTFAnimationSampler[];
    name?: any;
    extensions?: Record<string, any>;
    extras?: any;
}
/**
 * Metadata about the glTF asset.
 */
export interface GLTFAsset {
    /**
     * A copyright message suitable for display to credit the content creator.
     */
    copyright?: string;
    /**
     * Tool that generated this glTF model.  Useful for debugging.
     */
    generator?: string;
    /**
     * The glTF version that this asset targets.
     */
    version: string;
    /**
     * The minimum glTF version that this asset targets.
     */
    minVersion?: string;
    extensions?: Record<string, any>;
    extras?: any;
}
/**
 * A buffer points to binary geometry, animation, or skins.
 */
export interface GLTFBuffer {
    /**
     * The uri of the buffer.
     */
    uri?: string;
    /**
     * The length of the buffer in bytes.
     */
    byteLength: number;
    name?: any;
    extensions?: Record<string, any>;
    extras?: any;
}
/**
 * A view into a buffer generally representing a subset of the buffer.
 */
export interface GLTFBufferView {
    /**
     * The index of the buffer.
     */
    buffer: GLTFId;
    /**
     * The offset into the buffer in bytes.
     */
    byteOffset?: number;
    /**
     * The length of the bufferView in bytes.
     */
    byteLength: number;
    /**
     * The stride, in bytes.
     */
    byteStride?: number;
    /**
     * The target that the GPU buffer should be bound to.
     */
    target?: 34962 | 34963 | number;
    name?: any;
    extensions?: Record<string, any>;
    extras?: any;
}
// ...
export interface GLTFCamera {
    /**
     * An orthographic camera containing properties to create an orthographic projection matrix.
     */
    orthographic?: GLTFCameraOrthographic;
    /**
     * A perspective camera containing properties to create a perspective projection matrix.
     */
    perspective?: GLTFCameraPerspective;
    /**
     * Specifies if the camera uses a perspective or orthographic projection.
     */
    type: "perspective" | "orthographic" | string;
    name?: any;
    extensions?: Record<string, any>;
    extras?: any;
}
/**
 * Image data used to create a texture. Image can be referenced by URI or `bufferView` index. `mimeType` is required in the latter case.
 */
export interface GLTFImage {
    /**
     * The uri of the image.
     */
    uri?: string;
    /**
     * The image's MIME type.
     */
    mimeType?: "image/jpeg" | "image/png" | string;
    /**
     * The index of the bufferView that contains the image. Use this instead of the image's uri property.
     */
    bufferView?: GLTFId;
    name?: any;
    extensions?: Record<string, any>;
    extras?: any;
}
// ...
export interface GLTFMaterial {
    name?: any;
    extensions?: Record<string, any>;
    extras?: any;
    /**
     * A set of parameter values that are used to define the metallic-roughness material model from Physically-Based Rendering (PBR) methodology. When not specified, all the default values of `pbrMetallicRoughness` apply.
     */
    pbrMetallicRoughness?: GLTFMaterialPbrMetallicRoughness;
    /**
     * The normal map texture.
     */
    normalTexture?: GLTFMaterialNormalTextureInfo;
    /**
     * The occlusion map texture.
     */
    occlusionTexture?: GLTFMaterialOcclusionTextureInfo;
    /**
     * The emissive map texture.
     */
    emissiveTexture?: GLTFTextureInfo;
    /**
     * The emissive color of the material.
     */
    emissiveFactor?: number[];
    /**
     * The alpha rendering mode of the material.
     */
    alphaMode?: "OPAQUE" | "MASK" | "BLEND" | string;
    /**
     * The alpha cutoff value of the material.
     */
    alphaCutoff?: number;
    /**
     * Specifies whether the material is double sided.
     */
    doubleSided?: boolean;
}
/**
 * Geometry to be rendered with the given material.
 */
export interface GLTFMeshPrimitive {
    /**
     * A dictionary object, where each key corresponds to mesh attribute semantic and each value is the index of the accessor containing attribute's data.
     */
    attributes: Record<string, GLTFId>;
    /**
     * The index of the accessor that contains the indices.
     */
    indices?: GLTFId;
    /**
     * The index of the material to apply to this primitive when rendering.
     */
    material?: GLTFId;
    /**
     * The type of primitives to render.
     */
    mode?: 0 | 1 | 2 | 3 | 4 | 5 | 6 | number;
    /**
     * An array of Morph Targets, each  Morph Target is a dictionary mapping attributes (only `POSITION`, `NORMAL`, and `TANGENT` supported) to their deviations in the Morph Target.
     */
    targets?: Array<Record<string, GLTFId>>;
    extensions?: Record<string, any>;
    extras?: any;
}
/**
 * A set of primitives to be rendered.  A node can contain one mesh.  A node's transform places the mesh in the scene.
 */
export interface GLTFMesh {
    id?: string;
    /**
     * An array of primitives, each defining geometry to be rendered with a material.
     */
    primitives: GLTFMeshPrimitive[];
    /**
     * Array of weights to be applied to the Morph Targets.
     */
    weights?: number[];
    name?: any;
    extensions?: Record<string, any>;
    extras?: any;
}
// ...
export interface GLTFNode {
    /**
     * The index of the camera referenced by this node.
     */
    camera?: GLTFId;
    /**
     * The indices of this node's children.
     */
    children?: GLTFId[];
    /**
     * The index of the skin referenced by this node.
     */
    skin?: GLTFId;
    /**
     * A floating-point 4x4 transformation matrix stored in column-major order.
     */
    matrix?: number[];
    /**
     * The index of the mesh in this node.
     */
    mesh?: GLTFId;
    /**
     * The node's unit quaternion rotation in the order (x, y, z, w), where w is the scalar.
     */
    rotation?: number[];
    /**
     * The node's non-uniform scale, given as the scaling factors along the x, y, and z axes.
     */
    scale?: number[];
    /**
     * The node's translation along the x, y, and z axes.
     */
    translation?: number[];
    /**
     * The weights of the instantiated Morph Target. Number of elements must match number of Morph Targets of used mesh.
     */
    weights?: number[];
    name?: any;
    extensions?: Record<string, any>;
    extras?: any;
    userData?: any;
}
/**
 * Texture sampler properties for filtering and wrapping modes.
 */
export interface GLTFSampler {
    /**
     * Magnification filter.
     */
    magFilter?: 9728 | 9729 | number;
    /**
     * Minification filter.
     */
    minFilter?: 9728 | 9729 | 9984 | 9985 | 9986 | 9987 | number;
    /**
     * s wrapping mode.
     */
    wrapS?: 33071 | 33648 | 10497 | number;
    /**
     * t wrapping mode.
     */
    wrapT?: 33071 | 33648 | 10497 | number;
    name?: any;
    extensions?: Record<string, any>;
    extras?: any;
}
/**
 * The root nodes of a scene.
 */
export interface GLTFScene {
    /**
     * The indices of each root node.
     */
    nodes?: GLTFId[];
    name?: any;
    extensions?: Record<string, any>;
    extras?: any;
}
/**
 * Joints and matrices defining a skin.
 */
export interface GLTFSkin {
    id?: string;
    /**
     * The index of the accessor containing the floating-point 4x4 inverse-bind matrices.  The default is that each matrix is a 4x4 identity matrix, which implies that inverse-bind matrices were pre-applied.
     */
    inverseBindMatrices?: GLTFId;
    /**
     * The index of the node used as a skeleton root. When undefined, joints transforms resolve to scene root.
     */
    skeleton?: GLTFId;
    /** Indices of skeleton nodes, used as joints in this skin. */
    joints: GLTFId[];
    name?: any;
    extensions?: Record<string, any>;
    extras?: any;
}
/**
 * A texture and its sampler.
 */
export interface GLTFTexture {
    /** The index of the sampler used by this texture. When undefined, a sampler with repeat wrapping and auto filtering should be used. */
    sampler?: GLTFId;
    /** The index of the image used by this texture. */
    source?: GLTFId;
    name?: any;
    extensions?: Record<string, any>;
    extras?: any;
}
/**
 * The root object for a glTF asset.
 */
export interface GLTF {
    /**
     * Names of glTF extensions used somewhere in this asset.
     */
    extensionsUsed?: string[];
    /**
     * Names of glTF extensions required to properly load this asset.
     */
    extensionsRequired?: string[];
    /**
     * An array of accessors.
     */
    accessors?: GLTFAccessor[];
    /**
     * An array of keyframe animations.
     */
    animations?: GLTFAnimation[];
    /**
     * Metadata about the glTF asset.
     */
    asset: GLTFAsset;
    /**
     * An array of buffers.
     */
    buffers?: GLTFBuffer[];
    /**
     * An array of bufferViews.
     */
    bufferViews?: GLTFBufferView[];
    /**
     * An array of cameras.
     */
    cameras?: GLTFCamera[];
    /**
     * An array of images.
     */
    images?: GLTFImage[];
    /**
     * An array of materials.
     */
    materials?: GLTFMaterial[];
    /**
     * An array of meshes.
     */
    meshes?: GLTFMesh[];
    /**
     * An array of nodes.
     */
    nodes?: GLTFNode[];
    /**
     * An array of samplers.
     */
    samplers?: GLTFSampler[];
    /**
     * The index of the default scene.
     */
    scene?: GLTFId;
    /**
     * An array of scenes.
     */
    scenes?: GLTFScene[];
    /**
     * An array of skins.
     */
    skins?: GLTFSkin[];
    /**
     * An array of textures.
     */
    textures?: GLTFTexture[];
    extensions?: Record<string, unknown>;
    extras?: unknown;
    [k: string]: unknown;
}
// ...
export interface GLTF_KHR_draco_mesh_compression {
    bufferView: GLTFId;
    attributes: Record<string, number>;
    extras?: any;
}